Rename neon_local source file
