+++ control_plane/src/storage_controller.rs
